shopping/requirements.txt
asgiref==3.5.2
Babel==2.11.0
backports.zoneinfo==0.2.1
Django==3.2
django-cors-headers==3.13.0
django-extra-views==0.13.0
django-filter==22.1
django-haystack==3.2.1
django-oscar==3.1
django-oscar-accounts==3.0.1
django-oscar-api==3.1.2
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.4
phonenumbers==8.13.2
Pillow==9.3.0
purl==1.6
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
tzdata==2022.7
